[Spring] Throw exception when step definitions are annotated with component
When step definitions are annotated with @Component or other related
annotations they can be picked up by springs class path scanning. This
conflicts with cucumbers class path scanning and may result in
multiple bean definitions for the same class.

This problem is hard to understand and hard to trace. By making the
problem explicit and providing a clear instruction on how to resolve
this we can hopefully avoid future confusion.

The current implementation only checks the a subset of all annotations.
This will hopefully be sufficient.

Closes #1225